Bug 1951874: Discard the name section on invalid length. r=rhunt,jpages
It is possible for one of our custom sections to parse correctly even if
the declared length of the section does not agree with its contents. In
this case we should discard the results of the custom section parse.

Differential Revision: https://phabricator.services.mozilla.com/D241039

UltraBlame original commit: 062c45d9371d410ed9e89434d59a3ba5c689705c