Bug 1949695 - Make sure about:config is registered on Android. r=Gijs
In D240984, an attempt was made to re-enable `about:config` on Android
because it was removed by mistake.  That was done by reverting some
changes from D232785.

The intent was to keep `about:config` hidden in `about:about` on Android
but allow it to work unconditionally because GeckoView has a runtime
setting to control the presence/absence of this special page. I thought
not updating `components.conf` was precisely about that but that didn't
work in the end...

It is now clear to me that we should keep `'config'` in `components.conf`
and use a flag to hide the page in `about:about`. That's because this
config file is used at build time, therefore we always want `about:config`
available on Android, and it's up to the embedder to enable/disable it
with the runtime setting (that controls the `general.aboutConfig.enable`
pref under the hood).

Differential Revision: https://phabricator.services.mozilla.com/D242270

UltraBlame original commit: 99ae118679707da9c8229d028e6f237e5bd23373