Bug 1953446 - Part 2: Remove ObjectWeakMap class r=sfink
Previously ObjectWeakMap was a wrapper around a JSObject to JS::Value weakmap
with a slightly different API. This didn't make sense because we actually
support object to object maps directly, so storing the values as JS::Value is
pointless.

The patch removes this and replaces its uses with the standard weakmap
template. A 'maybeGet' method is added to get the value for a ket or a default
constructed Value if it is missing, since this is useful and frequently used.
Better name suggestions accepted.

Differential Revision: https://phabricator.services.mozilla.com/D241198

UltraBlame original commit: f86f9ae1d62ca12779d7240619d226d7274ac0c1