Bug 1954857 - Mirror calls to labeled_timing_distribution's sample APIs r=TravisLong
It's not consistent with the non-labeled case which converts to time_unit units
because labeled_timing_distribution doesn't know them.

But it's no more inconsistent than before, and unblocks further FOG Migration.

Hopefully in bug 1943453 we'll figure out a way to get around the weirdness.
And all else fails, Legacy Telemetry is set to be removed (along with GIFFT)
in the medium term.

Differential Revision: https://phabricator.services.mozilla.com/D242234

UltraBlame original commit: 10296d0618747f84b919200562289c5e4726d825