Bug 1953481 part 1 - avoid asserting when hitting timeout with TimeoutResult=warn r=dlp-reviewers,handyman
The code was expecting that once we get a response from the agent and
we are going to auto-acknowledge there should be an entry in
mUserActionIdToCanceledResponseMap. This was not the case if we were
waiting for a warn response from the user. Now we check for that case
and early return.

Differential Revision: https://phabricator.services.mozilla.com/D242376

UltraBlame original commit: e6de35b0e9d6aa4acf751f922b5fe9614667f3f1