Bug 1924461 - Rely on callers of GetNewOrUsedLaunchingBrowserProcess to handle a nullptr on shutdown. r=smaug
After bug 1955532 removed the last inconsistency we can rely on IsInOrBeyond(ShutdownPhase::AppShutdownConfirmed) to be sufficient to check such that the later assertions of a single process not being in shutdown are still meaningful.

Callers always had to react on nullptr, and in release we already did so here.

Differential Revision: https://phabricator.services.mozilla.com/D242649

UltraBlame original commit: 97d784aff636894cbb88f4bba23f08d5e1ca73dd