Bug 1955998 - Restore call to Theme::LookAndFeelChanged on init. r=dshin
This is not easy to test because this doesn't have any effect on
windows, and only startup is affected (so any ui.* pref that I could
tweak to write a reftest will "fix it").

I initially removed it to move it to nsXPLookAndFeel (thus making the
"refresh" and "init" code-paths more similar), but I think I forgot to
add it there, and it's not like it makes a big difference, so just put
it back where it was.

Differential Revision: https://phabricator.services.mozilla.com/D242746

UltraBlame original commit: 25b9102093f4c77c12b5709a921d390b62905983