Bug 1948647 - Do not dispatch the first-input event if there is a pending pointer down r=sefeng
The spec for dispatching the first-input event can be found here:
https://w3c.github.io/event-timing/#sec-dispatch-pending

Specifically step 6.2.2 handles this part of the code change.

The old code was following the spec, but some web-platform-tests were failing
because of it. After looking at the Chromium's codebase, I noticed that it
doesn't match the spec and it has this additional check. After adding it to our
code, it passes the tests as well. So it's clear that the spec doesn't match
the web platform tests or the Chromium implementation. It's more likely that
the spec needs to be updated.

Filed a spec bug here:
https://github.com/w3c/event-timing/issues/142

Differential Revision: https://phabricator.services.mozilla.com/D238470

UltraBlame original commit: a54d2326fdca6805038fb51a781ddf11712e09a6