Bug 1955697 - Remove nsDisplayCanvasBackgroundColor. r=mstange
We no longer need it, as now it behaves just like a solid color.

This causes an expectation change in an internal windowUtils.elementFromPoint
test, due to the HitTest implementation, but it's not observable to the web
(the web APIs return null when off-screen rather than the root), and no other
internal callers depend on it.

Differential Revision: https://phabricator.services.mozilla.com/D242445

UltraBlame original commit: bf3033ae1fe73d4384d9f3957f3efbfe4d36223a