Bug 1069183 - [devtools] Avoid too many tokenization in CSSCompleter. r=devtools-reviewers,ochameau.
`resolveState` now directly accepts an array of tokens so it won't do the tokenization,
and we take advantage of this in `traverseBackwards`, where we can pass the tokens
we computed at the top of `getInfoAt` and pop the last token from the stack until
we found the start of the "expression" we're interested in.

`browser_css_statemachine.js` is updated to test the 2 ways resolveState can be
called (with source + line + column, or with tokens).

Differential Revision: https://phabricator.services.mozilla.com/D241946

UltraBlame original commit: 661182d6133990709ba30a50541b5385a22a9fae