Bug 1954684 - Don't handle drag event before attaching window. r=geckoview-reviewers,ohall
I don't know why GeckoView crashes by this signature. But I found that
`onDragEvent` doesn't check whether `NativeProvider` is attached. It
might be a crash reason of this signature. So we should check it.

This occurs on start up, so I cannot write a test for this.

Differential Revision: https://phabricator.services.mozilla.com/D241939

UltraBlame original commit: a2759f08e124e7df785873dcec08d800da061964