Bug 1952914 - Part 2: Reduce complexity of checks in ContentParent::ValidatePrincipal, r=ckerschb,smaug
This refactors the code in ContentParent::ValidatePrincipal to make it do more
direct string operations on the RemoteType string. The idea is to reduce the
overhead of individual calls to ValidatePrincipal, now that it is being called
on more performance-sensitive codepaths.

In bug 1954597, I intend to move this code into ProcessIsolation.cpp, which
would also move it closer to the RemoteType building logic, hopefully reducing
the risk of this code getting out-of-sync with the process isolation logic.

Differential Revision: https://phabricator.services.mozilla.com/D242421

UltraBlame original commit: b49721f16330eae75d15354026c33d78d5d1d4f4