Bug 1956120 - Make HTMLFormElement::AddElementToTableInternal faster. r=smaug
The regression here was changed because of PositionIsBefore changing its
answer, but this is the same "mid-unbind position comparations get
weird".

Tweak the code a bit to reuse the comparator to share the cache, and
deal with `found` in one go.

Differential Revision: https://phabricator.services.mozilla.com/D242968

UltraBlame original commit: b6fb4070d81888bef42ca776a695788d79bcceec