Bug 1943622 - Move configure.sh to a simple key/value format r=glandium
Reuse the parse from confvars.sh to do so. This restricts the
expressiveness of the format as it limits it to variable declared as
confvar, but that's not a location where we actually want this kind of
customisation.

Differential Revision: https://phabricator.services.mozilla.com/D235770

UltraBlame original commit: 8ae4ed2e4692797d233539868dc00ac7a030b38e