Bug 1956046 - Don't clear CSD margin on tiling changes. r=stransky
This is meant to prevent spurious resize events when tiling / untiling
(because we still don't know the inner size allocate).

But if the conditions are just right, we might get stuck on an 'unknown'
margin:

 * The X server resizes us first, then changes the tiling state.
 * The GTK theme has the same decorations for tiled and untiled windows.

In that case, we don't get a configure, nor a container size-allocate
event, so the margin remains "unknown". Stick with the old margin.

Differential Revision: https://phabricator.services.mozilla.com/D243470

UltraBlame original commit: e0e88878d4b3bfaa70067dcd7050e8cfc8082790