Bug 1953167 - Part 3: Add JSHolderList container that doesn't use a hash table r=mccr8
This adds a container that supports the same operations as JSHolderMap but
doesn't use a hash map to look up the holder on add/remove. Instead it stores
the location of the entry in the holder itself.

I don't love this approach of adding extra parallel APIs so any suggestions for
improvement are welcome. This is kind of a request for feedback as much as a
review request.

Differential Revision: https://phabricator.services.mozilla.com/D241053

UltraBlame original commit: 553a0a5537408bd2dbc29ce4e7613e416c55d203