Bug 1957187 [wpt PR 51706] - [gap-decorations] Paint gap decorations for flex, a=testonly
Automatic update from web-platform-tests
[gap-decorations] Paint gap decorations for flex

This CL implements basic painting for Gap Decorations in flex.

In order to implement the rules, we had to add some data to the
`GapIntersections` as well as the `GapGeometry`. For flex, we use
`is_blocked_before_/after` to indicate whether the intersection borders
the edge of the container (is blocked by the edge).

This CL adds the `ContainerType` enum, and adds a member of this type to
`GapGeometry` which dictates for what type of container that
`GapGeometry` object is (i.e. flex).

We also add `inline_gap_size_`. In flex it refers to the gap between
flex items, and in grid it refers to the column gutter size.

For `block_gap_size_` in flex, it refers to the gap between flex lines
and in grid it refers to the row gutter size.

Bug: 357648037
Change-Id: Ifc65ba53049bff52388f57afda2471e3dce6adb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6367772
Commit-Queue: Javier Contreras <javierconmicrosoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajrmicrosoft.com>
Reviewed-by: Alison Maher <almahermicrosoft.com>
Reviewed-by: Kevin Babbitt <kbabbittmicrosoft.com>
Cr-Commit-Position: refs/heads/main{#1439763}

--

wpt-commits: 71ee931641c41f3f973711748b31c90f8049708c
wpt-pr: 51706

UltraBlame original commit: bf71b387244ed780063c843c7eca175d2744ed2f