Bug 1953025 - Add the check of reftest-wait in paintComplete(). r=webdriver-reviewers,emilio,whimboo
In `reftestWait()`, we await `paintComplete()` twice. The 2nd time we
force to wait only once because we are pretty sure the reftest-wait has
been removed. However, it is still possible to be blocked by the first call
of paintComplete() if there are too may MozAfterPaints. This is an
intermittent for view transitions when running the opacity animations.

Therefore, we add the check of removal of reftest-wait in
`paintComlete()`.

Differential Revision: https://phabricator.services.mozilla.com/D244282

UltraBlame original commit: 6679313b4d9b118a9a264f56bbf7703d53ae2292