Bug 1959151 - Stop hardcoding buildDir for Fenix ApkSizePlugin r=android-reviewers,nalexander
Recent changes to the Fenix buildDir in CI caused breakage of the ApkSizePlugin.
This updates the plugin to check buildDir variable instead of the projectDir so
that the plugin works regardless of where the build dir is.

Differential Revision: https://phabricator.services.mozilla.com/D244852

UltraBlame original commit: 0ae6a8477f03ac90825d0a8e6a06cb88351c71f1