Bug 1959305 - DoH pane undefined error in Privacy and Security when ParentalControlsServices disabled. r=settings-reviewers,Gijs
From: Sarah Jamie Lewis <sarahopenprivacy.ca>

When the parental controls service is disabled in a build, the DoH
settings now display the correct stauts when Increased or Max Protection
is enabled.

Previously, selecting either of these options would cause DoH to be
enabled, but the "Status" and "Provider" fields would not be properly
populated, due to a check on the gParentalControlsService causing an
error.

This check is now identical to the same check in DownloadIntegration.sys.mjs

Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: ma1 <giorgiomaone.net>

Differential Revision: https://phabricator.services.mozilla.com/D244850

UltraBlame original commit: 7c6dff2dd98cee69706e7f93913dc03ebdb889af