Bug 1954610: Fixed wpt transparent-text tests to verify 'last' entry rather than look at count r=sefeng
The assert was previously written with the assumption that only the
desired entry would be in the list. While the desired entry does need to
be the last one, it doesn't need to be the only one.

Differential Revision: https://phabricator.services.mozilla.com/D241892

UltraBlame original commit: 72d7a5668f5c55d812a69bb484a4e5704f19bbed