Bug 1956103 - Part 6: Migrate ScriptPreloader off of MonitorSingleWriter, r=arai
It appears that the only use of this being a MonitorSingleWriter is to avoid
holding `mSaveMonitor` to read `mCacheInvalidated` when starting a new content
process. So it seems somewhat likely that this could be simplified further,
removing the need for the special capabaility entirely.

Differential Revision: https://phabricator.services.mozilla.com/D242850

UltraBlame original commit: eb1cdf042dfc1b89571a12e812b66c07bf355818