Bug 1960055 - Make `mach bootstap` update out-of-date Android SDK directories. r=tcampbell
IIRC, this used to be hard due to the on-disk layout of the command
line tools packages.  But now, I don't think it is hard: we can try to
install the required version next to the outdated version without any
other destructive modifications, and then try to run it like normal.
If that works, great!  If not, back to the developer (for now -- see
Bug 1958986).

Differential Revision: https://phabricator.services.mozilla.com/D245290

UltraBlame original commit: f2c324291ab06e1e0978388275744432eb62cadb