Bug 1959927 - Fix perf tests that expect always a non-interruptible reflow. r=Gijs
The previous patch causes some performance tests to fail with unexpected
reflows. The issue is that I switched the logic to /avoid/ a flush,
which is surprising.

Upon investigation, as expected, it makes sense. Before my patch we had:

 * Refresh driver triggers interruptible reflow.
 * Resize observer handling triggers non-interruptible reflow (a no-op
   in browser.xhtml).

Now we only have the later, but it of course is no longer a no-op, so we
trigger the reflow observer machinery. That in turn dirties the root
frame, and causes subsequent good reflows to be misreported.

Don't dirty the frame tree from that code for those reflows, which we
ignored anyways.

Differential Revision: https://phabricator.services.mozilla.com/D245329

UltraBlame original commit: 31603a326303b80533bfb3262c262dd022533cbb