Bug 1959927 - Turn layout into a regular render phase. r=dshin
Run it in the right step of the rendering loop as per spec. Keep
the interruptible layout behavior by detecting whether we have any
frame with content-visibility: auto / ResizeObserver / etc around.

Also make the IntersectionObserver phase not unconditional, make it
conditional on needing layout or having scrolled.

I want to look a bit whether we can avoid flushing the whole browsing
context tree, but that is pre-existing so not touching it as part of
this work.

Differential Revision: https://phabricator.services.mozilla.com/D245214

UltraBlame original commit: 671d95927a0df5c3211af4526bb9566016ff03af