Bug 1957367 - QM: Introduce ClientDirectoryLockHandle as RAII wrapper for ClientDirectoryLock; r=dom-storage-reviewers,jstutte
This patch introduces ClientDirectoryLockHandle, a new RAII-style wrapper
around ClientDirectoryLock.

The handle ensures that the directory lock remains acquired for the lifetime
of the handle, and that appropriate cleanup occurs automatically when the
handle is destroyed. It is move-only and provides convenient access to the
underlying ClientDirectoryLock via operator overloads and a getter. It also
supports thread assertions through AssertIsOnOwningThread.

Currently, clients hold a RefPtr<ClientDirectoryLock> and must manually drop
the lock before nulling the RefPtr. This class is intended to make that
process safer and easier to follow. OpenClientDirectory will be updated in
a follow-up patch to return ClientDirectoryLockHandle.

Differential Revision: https://phabricator.services.mozilla.com/D243654

UltraBlame original commit: 050d50565d9fe20f5936e18dcc356ace8e9b0bc1