Bug 1438796 - Work around httpd deadlock in test_ext_contentscript_triggeringPrincipal.js r=zombie
As mentioned in bug 1408193, the test interaction with httpd results in
a deadlock, which causes the test to fail intermittently. I observed
this test failure consistently on debug builds.

This patch works around the issue by replacing the use of httpd with a
webRequest.onBeforeRequest listener to extract the relevant request body
from the request.

Along with that, fixes an issue in the ExtensionTestUtils test utility;
it modifies the input manifest, but because the test reuses the
EXTENSION_DATA input between test tasks, the special behavior of merging
host_permissions with permissions for MV2 causes a test failure in the
MV3 test (because in MV3, permissions may not contain host permissions).

Differential Revision: https://phabricator.services.mozilla.com/D240778

UltraBlame original commit: c4da70ccf2211095890a54a066e1964facd0f8ca