Bug 1792685 - Add pref to allow reverting to previous timing behavior r=zombie
Add pref to revert to previous timing behavior. This enables
users/developers to opt out of the new behavior, which may be useful to
confirm or rule out this bug as the cause of the issue.

This pref can be removed in a few cycles, once it has been established
that the changes did not cause undesired regressions.

Differential Revision: https://phabricator.services.mozilla.com/D241248

UltraBlame original commit: a2d827e489b6f6c71e683dd041f2c5627bb66c68