Bug 1957433 - QM: Add pref to disable origin access time updates in temporary storage for manual testing; r=dom-storage-reviewers,jstutte
Adds a new pref, dom.quotaManager.temporaryStorage.updateOriginAccessTime, to
allow disabling updates of origin access time in temporary storage.

This pref is intended for manual testing and debugging. The goal is to
temporarily disable access time tracking and confirm that our current test
suite does not fail, which highlights that we are missing test coverage for
this behavior.

While testing with the pref disabled, a few tests failed. This was due to the
fact that updating access time also sets the "accessed" flag, which causes
QuotaManager to perform a full origin scan on the next startup. Some tests
implicitly depend on that behavior and will be updated in a follow-up patch
to explicitly disable the quota info cache to avoid relying on that side effect.

The default value is true to preserve current behavior.

Differential Revision: https://phabricator.services.mozilla.com/D243765

UltraBlame original commit: df2294ed6d6f592e71a6a0dcb526afa50410ab6f