Bug 1960190 - Don't run field detection heurisitcs in identifyFieldsWhenFocused within dynamic form change timeout r=dimi,credential-management-reviewers
When receiving a focus event and handler.fillOnFormChangeData.isWithinDynamicFormChangeThreshold == true,
it's very likely that the focus event was dispatched by the site during the form change. In the case in which
the FormAutofillChild receives the "focus" event before the "form-changed" event we shoudln't process the field detection
in `identifyFieldsWhenFocused` because this updates the handler.fieldDetails and handler.form.elements. When we then handle
the `form-changed-event`, we abort early because the fieldDetails are already up-to-date and it seems that the fields haven't
changed, which prevents running a second autofill process on form change.

Differential Revision: https://phabricator.services.mozilla.com/D245365

UltraBlame original commit: 959dabe1a437ec08402d1b4a997d918726b439e0