Bug 1940581 - Remove Android ifdefs related to compositor scroll updates. r=hiro
In the future, we will need to send compositor scroll updates on all
platforms, since they will drive dynamic toolbar movement and we are
adding support for the dynamic toolbar in Responsive Design Mode.

We might as well remove the ifdefs now as that simplifies later patches
in this bug.

Note that we will not start actually sending NotifyCompositorScrollUpdate
IPC messages as a result of this patch, as PUICompositorController is only
instantiated on Android.

Differential Revision: https://phabricator.services.mozilla.com/D245523

UltraBlame original commit: 9db6a8d0d4c3f78510ec69f40170077265852bd6