Bug 1960811 - Part 6: Set a flag at the end of major sweeping r=sfink
A later patch will make it so we only take a lock to access the large buffer
map when we are sweeping. To minimise the time the lock is needed we can set a
flag on the background thread when major sweeping finishes and update our state
on the main thread when we detect that, rather than waiting for this to happen
as part of GC. This is the same approach as is used for minor sweeping.

Differential Revision: https://phabricator.services.mozilla.com/D245867

UltraBlame original commit: e7b2a7c60c351ecd96bc9e0cf917b309b7ab2d57