Bug 1961080 - Add Prio3Histogram VDAF to DAPTelemetrySender r=simonf
The Prio3Histogram VDAF is similar to a one-hot encoding applied to a sumvec
VDAF, but an integer measurement is used to specify the non-zero entry instead
of a vector. This has a more optimal encoding and better characteristics if
differential privacy is applied to the task.

Differential Revision: https://phabricator.services.mozilla.com/D245848

UltraBlame original commit: 15c1ca1556154d5a757faf8d1beec8153a002ab6