Bug 1948633 - part 3 - Render pageOrigin from the state r=android-reviewers,android-l10n-reviewers,delphine,tchoh
Because of a Jetpack Compose bug which prevents properly anchoring
Dropdowns to the bottom of the screen in order to support showing
the long press menu from a bottom toolbar I had to use a View for
the origin instead of a Composable.

Implementation is to change after Google fixes the Compose bug.

Differential Revision: https://phabricator.services.mozilla.com/D245953

UltraBlame original commit: 4844ed893da8ad89469d30b049edbaabeffe060e