Bug 1962140 - Update the match rule of the Asana Zendesk shim. r=webcompat-reviewers
The patch updates the match rule of the contentScript of the Asana
Zendesk shim. The Zendesk page navigates between pages using History
API, so it won't trigger an actual loading. So, the existing shim won't
get injected.

We update the match rule to let the shim to be injected at the very
beginning when loading a Zendesk page to solve the issue.

Differential Revision: https://phabricator.services.mozilla.com/D246420

UltraBlame original commit: 1ae66a70a4766a0ca1a7567587dc12ba462d07b8