Bug 1962688 - Fix tab drop zone calculation to account for window decorations correctly. r=tabbrowser-reviewers,nsharpley
This is more visible on Linux because the decorations there are bigger, but I
believe this also fixes issues on windows specially if you have the titlebar
checkbox on.

You need to use mozInnerScreen* coords in order to get the screen coordinates
of the client area.

Differential Revision: https://phabricator.services.mozilla.com/D246829

UltraBlame original commit: ea19d60a97e9b03b595fe82ded963092d13c8cae