Bug 1958280 part 5 - Add iterators for LInstruction outputs and temps. r=iain
The boilerplate for this (including skipping `BogusTemp` definitions) is a bit tedious
so add an iterator for this.

The term 'definition' is used for both outputs and temps so use `OutputIter` for outputs
to be more specific.

Differential Revision: https://phabricator.services.mozilla.com/D246442

UltraBlame original commit: 8d60ccbe17b34259b2ba5cf78b89b4b83c0b58a9