Bug 1958280 part 9 - Fix minor regression in addLiveRegistersForRange. r=jseward
This fixes a regression from bug 1938317 part 3 that I noticed while reading the code.

We'd incorrectly mark an instruction's output register as a live-register and
save/restore it while we didn't before that change.

We can't easily check for this in the regalloc verifier because the same register
could also be added to the safepoint's live-regs for a different vreg that was live
at the start of the instruction.

Differential Revision: https://phabricator.services.mozilla.com/D246446

UltraBlame original commit: 621f05dd87a6d27985f994d175b01c05faa0c126