Bug 1948315: rename `release_version_bump` transform to reflect reality r=releng-reviewers,taskgraph-reviewers,ahal
This transform may have been specific to `release_version_bump` at some point, but it's certainly not at this point, so rename it.

It almost seems silly for it to exist at all, but I was unable to replace it with `task-context`, so this rename will have to do.

Differential Revision: https://phabricator.services.mozilla.com/D246817

UltraBlame original commit: e83c576b567fe4e0c885b21e29893f87a41cf9f0